Count packets which haven't received a response after a timeout.
This allows us to count "timed out" packets where we no longer expect to
get a response back - this provides a way of measuring packet loss which
doesn't fluctuate - e.g. for smokeping-prober there's a slightly
annoying drawback in that if you sample when a packet has been sent but
before its response has come back, then you'll always see "some" loss.

To avoid this, we count the number of packets we expect to have seen.

TODO: unbreak this in the presence of duplicate packets